+++ templates/inferno-app/src/App.js
@@ -1,7 +1,6 @@
 import './App.css'
 
-import Inferno from 'inferno'
-import Component from 'inferno-component'
+import {Component} from 'inferno'
 
 class App extends Component {
   render() {